requirements.txt
torch>=2.11.0
transformers>=4.57.0
datasets>=4.6.0
peft>=0.18.0
trl>=1.3.0
accelerate>=1.13.0
bitsandbytes>=0.49.0
deepspeed>=0.18.0
wandb
autoawq
evaluate
sentencepiece
